+++ spec/shelly/cli/deploy_spec.rb
@@ -113,6 +113,8 @@ describe Shelly::CLI::Deploy do
       $stdout.should_receive(:puts).with("delayed jobs")
       $stdout.should_receive(:puts).with(green "Starting sidekiq")
       $stdout.should_receive(:puts).with("sidekiq workers")
+      $stdout.should_receive(:puts).with(green "Starting clockwork")
+      $stdout.should_receive(:puts).with("clockwork output")
       $stdout.should_receive(:puts).with(green "Starting thin")
       $stdout.should_receive(:puts).with("thins up and running")
       $stdout.should_receive(:puts).with(green "Starting puma")
@@ -135,7 +137,7 @@ describe Shelly::CLI::Deploy do
         "after_restart" => "after restart hook", "before_restart" => "before restart hook",
         "before_symlink" => "before symlink hook", "before_migrate" => "before migrate hook",
         "after_successful_deploy_hook" => "after successful deploy hook",
-        "on_restart" => "on restart hook"
+        "on_restart" => "on restart hook", "clockwork" => "clockwork output"
       }
     end
   end